Deleted extraneous files
There were several files in different folders that were just blank so I deleted them. Additionally, in the fish module we initially used rData and then switched to csv so I deleted the rData files. In mod 3 there was a biomass_old.csv file that was not used (presumable replaced with biomass,csv) so I deleted that as well. I also deleted sites.csv because it had the exact same data as bison_sites but was not used in the analysis.